requirements.txt
python-telegram-bot[job-queue]==21.0.1
requests==2.31.0
python-dotenv==1.0.1
matplotlib==3.10.1
pytz==2024.1
